third_party/WebKit/LayoutTests/editing/selection/modify_move/move_forward_line_small_line.html

<!doctype html>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script src="../../assert_selection.js"></script>
<script>
test(() => assert_selection(
        [
            '<p contenteditable style="line-height: 1em; white-space: pre-line">',
                '|foo\nbar',
            '</p>',
        ].join(''),
        selection => selection.modify('move', 'forward', 'line'),
        [
            '<p contenteditable style="line-height: 1em; white-space: pre-line">',
                'foo\n|bar',
            '</p>',
        ].join('')),
        'Forward line on small lines');
</script>
